Moved job title to its own folder and changed paths
